+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 2. [ohos]简化实现'修复连续打开同一个弹窗再回退时页面卡死的问题'
 3. [ios]修复打开一个新的container时会执行多次surfaceUpdated的问题
 4. Revert "[ios]修复打开一个新的container时会执行多次surfaceUpdated的问题" Reason: 该修改存在bad case，需要换一种实现方式
+5. [ios]重新实现“修复打开一个新的container时会执行多次surfaceUpdated的问题”
 
 ## 4.6.2
 1. 更新README以及FAQ